\name{LungCancerFastqFiles}
\alias{LungCancerFastqFiles}
\title{
  Get the Fastq paths
}
\description{
  Returns a character vector of file paths to the demo Fastq files.
}
\usage{
LungCancerFastqFiles()
}
\value{
  A character vector, named according to
  \dQuote{H[1993/2073].[first/last]}.
}
\author{
  Michael Lawrence
}

\examples{
LungCancerFastqFiles()
}
