\name{TP53Genome-package}
\alias{TP53Genome-package}
\alias{TP53Genome}
\docType{package}
\title{
  Genomic Sequence of the TP53 Gene Plus a 1-Megabase Region on Each
  Side of the Gene
}
\description{
  This package was created to use in examples and test sets for the
  gmapR and VariationTools packages.
}
\details{
\tabular{ll}{
Package: \tab TP53Genome\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-05\cr
License: \tab Artistic-2.0\cr
}
By calling data(p53Genome), users will have access to a GmapGenome
object for the TP53 genome.
}
\author{
  Cory Barr
  
Maintainer: Cory Barr <barr.cory@gene.com>
}
\examples{
data(p53Genome)
}
